+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG](http://keepachangelog.com/) for how to update this file. This project
 adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Add `referrer` and `userAgent` to notice object to support filtering
 
 ## [2.3.0-beta.0] - 2020-07-28
 ### Added